Extract base s3 uploader for export
