+++ ydb/core/backup/s3/extstorage_usage_config.cpp
@@ -2,12 +2,12 @@
 #include <ydb/core/wrappers/s3_storage_config.h>
 
 #ifndef KIKIMR_DISABLE_S3_OPS
-namespace NKikimr::NDataShard {
+namespace NKikimr::NBackup::NS3 {
 
 Aws::S3::Model::StorageClass TS3Settings::GetStorageClass() const {
     return NKikimr::NWrappers::NExternalStorage::TS3ExternalStorageConfig::ConvertStorageClass(StorageClass);
 }
 
-}
+} // namespace NKikimr::NBackup::NS3
 
 #endif // KIKIMR_DISABLE_S3_OPS
