+++ ydb/core/backup/ya.make
@@ -1,4 +1,6 @@
 RECURSE(
+    common
     controller
     impl
+    s3
 )
