+++ ydb/core/tx/datashard/ya.make
@@ -4,7 +4,6 @@ SRCS(
     alter_cdc_stream_unit.cpp
     alter_table_unit.cpp
     backup_restore_common.cpp
-    backup_restore_traits.cpp
     backup_unit.cpp
     build_and_wait_dependencies_unit.cpp
     build_data_tx_out_rs_unit.cpp
@@ -220,7 +219,6 @@ SRCS(
     wait_for_stream_clearance_unit.cpp
 )
 
-GENERATE_ENUM_SERIALIZATION(backup_restore_traits.h)
 GENERATE_ENUM_SERIALIZATION(change_exchange.h)
 GENERATE_ENUM_SERIALIZATION(datashard.h)
 GENERATE_ENUM_SERIALIZATION(datashard_active_transaction.h)
@@ -253,6 +251,8 @@ PEERDIR(
     library/cpp/l1_distance
     library/cpp/l2_distance
     ydb/core/actorlib_impl
+    ydb/core/backup/common
+    ydb/core/backup/s3
     ydb/core/base
     ydb/core/change_exchange
     ydb/core/engine
@@ -293,7 +293,6 @@ ELSE()
         export_s3_buffer_raw.cpp
         export_s3_buffer_zstd.cpp
         export_s3_uploader.cpp
-        extstorage_usage_config.cpp
         import_s3.cpp
     )
 ENDIF()
